requirements.txt
anyio==3.5.0
appnope==0.1.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.3.2
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.3.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
cramjam==2.5.0
cycler==0.11.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.3
fastjsonschema==2.15.3
fastparquet==0.7.2
filelock==3.6.0
flake8==4.0.1
fonttools==4.29.1
fsspec==2021.11.1
identify==2.4.12
idna==3.3
imageio==2.5.0
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.2
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.10.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.0
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyter-server==1.16.0
jupyterlab==3.3.2
jupyterlab-code-formatter==1.4.10
jupyterlab-pygments==0.1.2
jupyterlab-server==2.12.0
jupyterlab-widgets==1.0.2
kaggle==1.5.5
kiwisolver==1.3.2
lazy-object-proxy==1.4.2
lightgbm==3.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
mccabe==0.6.1
mistune==0.8.4
-e git+git@github.com:Katsuya-Ishiyama/ml_framework.git@6f92e2bb4df9afb4a4156bfbd79d92a8bfbc3a64#egg=ml_framework
mypy==0.942
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.9
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio==1.5.4
networkx==2.3
nodeenv==1.6.0
notebook==6.4.7
notebook-shim==0.1.0
numpy==1.19.5
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.1
plotly==5.5.0
pluggy==1.0.0
pre-commit==2.18.1
prometheus-client==0.12.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
py==1.11.0
pyarrow==6.0.1
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.4.3
pyparsing==3.0.6
pyrsistent==0.18.0
pytest==6.2.5
python-dateutil==2.8.2
python-slugify==3.0.3
pytz==2021.3
PyWavelets==1.0.3
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.0
requests==2.27.1
scikit-image==0.15.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.9.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.2
threadpoolctl==3.1.0
thrift==0.15.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.36.0
traitlets==5.1.1
typed-ast==1.5.2
typing-extensions==4.0.0
urllib3==1.26.9
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
widgetsnbextension==3.5.2
wrapt==1.11.2
xgboost==0.90
zipp==3.6.0
